Change hit_by_cob to hit_by_ash and support all ashes

+++ lib/object/projectile.h
@@ -36,7 +36,7 @@ class zombie;
 
 class projectile {
 public:
-    uint64_t from_plant;
+    int from_plant;
     projectile_type type;
     projectile_motion_type motion_type;
